+++ compiler/components/DNASignGraphic.py
@@ -34,7 +34,7 @@ class DNASignGraphic(DNANode.DNANode):
         packer.pack('code', self.code, SHORT_STRING)
 
         for component in self.color:
-            packer.pack('color', int(component * 255), UINT16)
+            packer.pack('color', int(component * 255), UINT8)
 
         packer.pack('width', int(self.width * 100), INT32)
         packer.pack('height', int(self.height * 100), INT32)
