+++ compiler/components/DNAWindows.py
@@ -28,7 +28,7 @@ class DNAWindows(DNAGroup.DNAGroup):
         packer.pack('code', self.code, SHORT_STRING)
 
         for component in self.color:
-            packer.pack('color', int(component * 255), UINT16)
+            packer.pack('color', int(component * 255), UINT8)
 
         packer.pack('window count', self.windowCount, UINT8)
         return packer
